feat: Adds customer authentication

+++ src/main/java/com/vodafone/controller/CustomerController.java
@@ -3,6 +3,7 @@ package com.vodafone.controller;
 import com.vodafone.model.*;
 import com.vodafone.model.dto.CreateUser;
 import com.vodafone.service.*;
+import com.vodafone.validators.UserAuthorizer;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -27,27 +28,39 @@ public class CustomerController {
     private ProductService productService;
     private CartService cartService;
     private SendEmailService sendEmailService;
+    private UserAuthorizer userAuthorizer;
 
     // Home
     @GetMapping("home.htm")
-    public String home(Model model, @RequestParam(required = false) String category) {
-        List<Product> products;
-        if (category != null)
-            products = this.productService.getByCategory(category);
-        else
-            products = this.productService.getAll();
-        model.addAttribute("products", products);
-        return "/customer/shared/home";
+    public String home(HttpSession session, Model model, @RequestParam(required = false) String category) {
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            List<Product> products;
+            if (category != null)
+                products = this.productService.getByCategory(category);
+            else
+                products = this.productService.getAll();
+            model.addAttribute("products", products);
+            return "/customer/shared/home";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     @GetMapping("/search/home.htm")
-    public String search(Model model, @RequestParam(required = false) String category, @RequestParam(required = false) String name) {
-        List<Product> products = new ArrayList<>(this.productService.getByCategory(category));
-        Product selectedProduct = this.productService.getByName(name);
-        if (selectedProduct != null)
-            products.add(selectedProduct);
-        model.addAttribute("products", products);
-        return "/customer/shared/home";
+    public String search(HttpSession session,Model model, @RequestParam(required = false) String category,
+                         @RequestParam(required = false) String name) {
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            List<Product> products = new ArrayList<>(this.productService.getByCategory(category));
+            Product selectedProduct = this.productService.getByName(name);
+            if (selectedProduct != null)
+                products.add(selectedProduct);
+            model.addAttribute("products", products);
+            return "/customer/shared/home";
+        }
+        else {
+            return "redirect:/login.htm";
+        }
     }
 
     @PostMapping
@@ -76,30 +89,46 @@ public class CustomerController {
     }
 
     @GetMapping("reset.htm")
-    public String resetPasswordLoader(Model model) {
-        model.addAttribute("resetUser", new CreateUser());
-        return "resetPassword";
+    public String resetPasswordLoader(HttpSession session,Model model) {
+        if(userAuthorizer.customerExists(session)) {
+            model.addAttribute("resetUser", new CreateUser());
+            return "resetPassword";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     @PostMapping("reset.htm")
-    public String resetPassword(@Valid @ModelAttribute("resetUser") CreateUser user, BindingResult bindingResult, HttpSession session) {
-        if (bindingResult.hasErrors()) {
-            Map<String, Object> modelBind = bindingResult.getModel();
-            System.out.println(modelBind);
-            return "registration";
+    public String resetPassword(@Valid @ModelAttribute("resetUser") CreateUser user,
+                                BindingResult bindingResult, HttpSession session) {
+        if(userAuthorizer.customerExists(session)) {
+            if (bindingResult.hasErrors()) {
+                Map<String, Object> modelBind = bindingResult.getModel();
+                System.out.println(modelBind);
+                return "registration";
+            }
+            System.out.println(user);
+            if (customerService.resetPassword(session.getAttribute("email").toString(), user.getPassword()))
+                return "login";
+            return "resetPassword";
         }
-        System.out.println(user);
-        if (customerService.resetPassword(session.getAttribute("email").toString(), user.getPassword()))
-            return "login";
-        return "resetPassword";
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
 
     @GetMapping("/products/{id}/details.htm")
-    public String viewProductDetails(Model model, @PathVariable Long id) {
-        Product product = this.productService.get(id);
-        model.addAttribute("product", product);
-        return "/customer/product/detail";
+    public String viewProductDetails(HttpSession session,Model model, @PathVariable Long id) {
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Product product = this.productService.get(id);
+            model.addAttribute("product", product);
+            return "/customer/product/detail";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     @GetMapping("products/rate")
@@ -135,43 +164,48 @@ public class CustomerController {
     //edit in frontend
     @GetMapping("/orders.htm")
     public String getCustomerOrders(HttpSession session, Model model) {
-        Long customerId = (long) session.getAttribute("id");
-        List<Order> orders = orderService.getByCustomerId(customerId);
-        model.addAttribute("orders", orders);
-        return "/customer/shared/orders";
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            List<Order> orders = orderService.getByCustomerId(customerId);
+            model.addAttribute("orders", orders);
+            return "/customer/shared/orders";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
-
-    /*@GetMapping("{customerId}/finalOrder")
-    public String showFinalOrder(@PathVariable Long customerId) {
-        Cart customerCart = customerService.get(customerId).getCart();
-        Order finalOrder = cartService.showFinalOrder(customerCart.getId());
-        return null;
-    }*/
-    //change in front end
     @PostMapping("/submitOrder.htm")
     @ResponseBody
     public String submitFinalOrder(HttpSession session) {
-        Long customerId = (long) session.getAttribute("id");
-        Cart customerCart = customerService.get(customerId).getCart();
-        Order submittedOrder = cartService.submitFinalOrder(customerCart.getId());
-        boolean created = orderService.create(submittedOrder);
-        if (created)
-            return "true";
-        //todo redirect to error page
-        return "false";
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            Cart customerCart = customerService.get(customerId).getCart();
+            Order submittedOrder = cartService.submitFinalOrder(customerCart.getId());
+            boolean created = orderService.create(submittedOrder);
+            if (created)
+                return "true";
+            return "false";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
-    //change in front end
     @GetMapping("showCart.htm")
     public String showCustomerCart(Model model, HttpSession session) {
-        Long customerId = (long) session.getAttribute("id");
-        Cart customerCart = customerService.get(customerId).getCart();
-        List<CartItem> cartItems = customerCart.getItems();
-        double totalCartPrice = cartItems.stream().mapToDouble(CartItem::getTotal).sum();
-        model.addAttribute("items", cartItems);
-        model.addAttribute("orderTotal", totalCartPrice);
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            Cart customerCart = customerService.get(customerId).getCart();
+            List<CartItem> cartItems = customerCart.getItems();
+            double totalCartPrice = cartItems.stream().mapToDouble(CartItem::getTotal).sum();
+            model.addAttribute("items", cartItems);
+            model.addAttribute("orderTotal", totalCartPrice);
 
-        return "/customer/shared/cart";
+            return "/customer/shared/cart";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     //change in front end
@@ -179,42 +213,45 @@ public class CustomerController {
     @ResponseBody
     public String addItemToCart(HttpSession session, @RequestParam Long itemId,
                                 @RequestParam int quantity) {
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            Cart customerCart = customerService.get(customerId).getCart();
+            Product product = productService.get(itemId);
 
-        Long customerId = (long) session.getAttribute("id");
-        Cart customerCart = customerService.get(customerId).getCart();
-        Product product = productService.get(itemId);
-
-        CartItem cartItem = new CartItem(quantity, product, customerCart);
-        boolean added = cartService.addItem(customerCart.getId(), cartItem);
-        if (added)
-            return "true";
-        return "false";
+            CartItem cartItem = new CartItem(quantity, product, customerCart);
+            boolean added = cartService.addItem(customerCart.getId(), cartItem);
+            if (added)
+                return "true";
+            return "false";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     //change in front end
     @DeleteMapping("showCart.htm")
     @ResponseBody
     public String removeItemFromCart(HttpSession session, @RequestParam Long itemId) {
-        Long customerId = (long) session.getAttribute("id");
-        Cart customerCart = customerService.get(customerId).getCart();
-        boolean deleted = cartService.removeItem(customerCart.getId(), itemId);
-        if (deleted)
-            return "true";
-        return "false";
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            Cart customerCart = customerService.get(customerId).getCart();
+            boolean deleted = cartService.removeItem(customerCart.getId(), itemId);
+            if (deleted)
+                return "true";
+            return "false";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
-
-    /*@PutMapping("{customerId}/cart/clear")
-    public String clearCustomerCart(@PathVariable Long customerId) {
-        Cart customerCart = customerService.get(customerId).getCart();
-        cartService.clearCart(customerCart.getId());
-        return null;
-    }*/
 
 
     @GetMapping("registration.htm")
     public String registration(Model model) {
-        model.addAttribute("customerDTO", new Customer());
-        return "registration";
+            model.addAttribute("customerDTO", new Customer());
+            return "registration";
+
     }
 
     //todo: change customer param type to CreateUser DTO
@@ -244,79 +281,101 @@ public class CustomerController {
 
     @GetMapping("resend.htm")
     public String resendOtp(HttpSession session) {
-        Customer selectedCustomer = this.customerService.getByUserName(session.getAttribute("username").toString());
-        String otp = this.sendEmailService.getRandom();
-        selectedCustomer.setCode(otp);
-        session.setAttribute("verificationCode", otp);
-        this.customerService.update(selectedCustomer.getId(),selectedCustomer);
-        if (sendEmailService.sendEmail(selectedCustomer, EmailType.ACTIVATION, session)) {
+        if(userAuthorizer.customerExists(session)) {
+            Customer selectedCustomer = this.customerService.getByUserName(session.getAttribute("username").toString());
+            String otp = this.sendEmailService.getRandom();
+            selectedCustomer.setCode(otp);
+            session.setAttribute("verificationCode", otp);
+            this.customerService.update(selectedCustomer.getId(), selectedCustomer);
+            if (sendEmailService.sendEmail(selectedCustomer, EmailType.ACTIVATION, session)) {
 
-            return "redirect:/customer/verify.htm";
-        } else {
-            return "registration";
+                return "redirect:/customer/verify.htm";
+            } else {
+                return "registration";
+            }
+        }
+        else{
+            return "redirect:/login.htm";
         }
     }
 
     @GetMapping("/verify.htm")
     public String verify(Model model, HttpSession session) {
-        model.addAttribute("customer", new Customer());
+        if(userAuthorizer.customerExists(session)) {
+            model.addAttribute("customer", new Customer());
 
-        final ScheduledThreadPoolExecutor executor = new ScheduledThreadPoolExecutor(1);
-        executor.schedule(() -> {
-            customerService.expireOtp(session.getAttribute("username").toString());
-        }, 1, TimeUnit.MINUTES);
+            final ScheduledThreadPoolExecutor executor = new ScheduledThreadPoolExecutor(1);
+            executor.schedule(() -> {
+                customerService.expireOtp(session.getAttribute("username").toString());
+            }, 1, TimeUnit.MINUTES);
 
-        return "verify";
+            return "verify";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
 
     @PostMapping("verify.htm")
     public String verifyCustomer(@Valid @ModelAttribute("customer") Customer customer, BindingResult bindingResult, HttpSession session) {
-        if (bindingResult.hasErrors()) {
-            Map<String, Object> modelBind = bindingResult.getModel();
-            System.out.println(modelBind);
-            return "verify";
-        }
-        Customer customer1 = customerService.getByMail((String) session.getAttribute("email"));
-        if (customer1 == null) {
-            //todo: display email not found error
-            return "404";
-        } else {
-            if (customer1.getCode().equals(customer.getCode())) {
-                System.out.println(customer1.getEmail());
-                System.out.println("updated " + customerService.updateStatusActivated(customer1.getEmail()));
-                customerService.updateStatusActivated(customer1.getEmail());
-                return "redirect:/customer/home.htm";
-            } else {
-                return "404";
+        if(userAuthorizer.customerExists(session)) {
+            if (bindingResult.hasErrors()) {
+                Map<String, Object> modelBind = bindingResult.getModel();
+                System.out.println(modelBind);
+                return "verify";
             }
+            Customer customer1 = customerService.getByMail((String) session.getAttribute("email"));
+            if (customer1 == null) {
+                //todo: display email not found error
+                return "404";
+            } else {
+                if (customer1.getCode().equals(customer.getCode())) {
+                    System.out.println(customer1.getEmail());
+                    System.out.println("updated " + customerService.updateStatusActivated(customer1.getEmail()));
+                    customerService.updateStatusActivated(customer1.getEmail());
+                    return "redirect:/customer/home.htm";
+                } else {
+                    return "404";
+                }
+            }
+        }
+        else{
+            return "redirect:/login.htm";
         }
     }
 
 
     @PutMapping("/increment")
     @ResponseBody
-
     public String incrementProductQuantity(HttpSession session, @RequestParam Long productId) {
-        //retrieve cartId from session
-        Long customerId = (long) session.getAttribute("id");
-        Long cartId = customerService.get(customerId).getCart().getId();
-        int newQuantity = cartService.incrementProductQuantity(cartId, productId, 1);
-        if (newQuantity > 0)
-
-            return "true";
-        return "false";
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            //retrieve cartId from session
+            Long customerId = (long) session.getAttribute("id");
+            Long cartId = customerService.get(customerId).getCart().getId();
+            int newQuantity = cartService.incrementProductQuantity(cartId, productId, 1);
+            if (newQuantity > 0)
+                return "true";
+            return "false";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
 
     @PutMapping("/decrement")
     @ResponseBody
     public String decrementProductQuantity(HttpSession session, @RequestParam Long productId) {
-        Long customerId = (long) session.getAttribute("id");
-        Long cartId = customerService.get(customerId).getCart().getId();
-        int newQuantity = cartService.decrementProductQuantity(cartId, productId);
-        if (newQuantity >= 0)
-            return "true";
-        return "false";
+        if(userAuthorizer.isActivatedCustomer(session)) {
+            Long customerId = (long) session.getAttribute("id");
+            Long cartId = customerService.get(customerId).getCart().getId();
+            int newQuantity = cartService.decrementProductQuantity(cartId, productId);
+            if (newQuantity >= 0)
+                return "true";
+            return "false";
+        }
+        else{
+            return "redirect:/login.htm";
+        }
     }
-
 }
